Put all price-related code into src/simulation/prices.rs
